[AV1e Gacc] Merge Latest AV1 GACC plugin to master
Added Unit tests.
Using mfxExtAV1Param.
Removed duplicate options from mfx_transcoder.

Issue: N\A
Tests: Internal AMT testing with standard tests

+++ _studio/mfx_lib/fs_det/src/Cleanup.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2014-2018 Intel Corporation
+// Copyright (c) 2014-2019 Intel Corporation
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -18,18 +18,18 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 /********************************************************************************
- * 
+ *
  * File: Cleanup.cpp
  *
  * Structures and routines for binary mask cleanup (as postprocessing)
- * 
+ *
  ********************************************************************************/
 #include <stdio.h>
 #include <limits.h>
 #include "Cleanup.h"
 
 //Pop element from stack
-int pop(int *x, int *y, int pitch, int stack[], int& stackptr) 
+int pop(int *x, int *y, int pitch, int stack[], int& stackptr)
 {
     if(stackptr > 0) {
         int p = stack[stackptr];
@@ -39,10 +39,10 @@ int pop(int *x, int *y, int pitch, int stack[], int& stackptr)
         return 1;
     }
     else return 0;
-}    
+}
 
 //Push element to stack
-int push(int x, int y, int pitch, int stack[], int& stackptr, int MAX_BLK_RESOLUTION) 
+int push(int x, int y, int pitch, int stack[], int& stackptr, int MAX_BLK_RESOLUTION)
 {
     if(stackptr < MAX_BLK_RESOLUTION - 1) {
         stackptr++;
@@ -53,14 +53,14 @@ int push(int x, int y, int pitch, int stack[], int& stackptr, int MAX_BLK_RESOLU
 }
 
 //Empty stack
-void empty_stack(int pitch, int stack[], int& stackptr) 
+void empty_stack(int pitch, int stack[], int& stackptr)
 {
-    int x, y; 
+    int x, y;
     while(pop(&x, &y, pitch, stack, stackptr));
 }
 
 //Non-recursively flood-fills a 4-connected component in a block
-void flood_fill4(int x, int y, int nc, int oc, int w, int h, BYTE *frm, int stack[], int& stackptr, int MAX_BLK_RESOLUTION) 
+void flood_fill4(int x, int y, int nc, int oc, int w, int h, BYTE *frm, int stack[], int& stackptr, int MAX_BLK_RESOLUTION)
 {
     if(nc == oc) return; //avoid infinite loop
     empty_stack(w, stack, stackptr);
@@ -68,7 +68,7 @@ void flood_fill4(int x, int y, int nc, int oc, int w, int h, BYTE *frm, int stac
     if(!push(x, y, w, stack, stackptr, MAX_BLK_RESOLUTION)) return;
 
     while(pop(&x, &y, w, stack, stackptr)) {
-        frm[y*w + x] = nc;
+        frm[y*w + x] = (BYTE) nc;
         //4-way filling
         if (x+1 <w) {
             if (frm[y*w + (x+1)]==oc && frm[y*w + (x+1)]!=nc) {
@@ -80,7 +80,7 @@ void flood_fill4(int x, int y, int nc, int oc, int w, int h, BYTE *frm, int stac
                 if(!push(x-1, y, w, stack, stackptr, MAX_BLK_RESOLUTION)) return;
             }
         }
-        if (y+1 <h) { 
+        if (y+1 <h) {
             if (frm[(y+1)*w + x]==oc && frm[(y+1)*w + x]!=nc) {
                 if(!push(x, y+1, w, stack, stackptr, MAX_BLK_RESOLUTION)) return;
             }
@@ -94,7 +94,7 @@ void flood_fill4(int x, int y, int nc, int oc, int w, int h, BYTE *frm, int stac
 }
 
 //Non-recursively flood-fills a 4-connected component in a block and collects the component's size
-void flood_fill4_sz(int x, int y, int nc, int oc, int w, int h, BYTE *frm, int *sz, int stack[], int& stackptr, int MAX_BLK_RESOLUTION) 
+void flood_fill4_sz(int x, int y, int nc, int oc, int w, int h, BYTE *frm, int *sz, int stack[], int& stackptr, int MAX_BLK_RESOLUTION)
 {
     (*sz) = 0;
     if(nc == oc) return; //avoid infinite loop
@@ -105,7 +105,7 @@ void flood_fill4_sz(int x, int y, int nc, int oc, int w, int h, BYTE *frm, int *
     while(pop(&x, &y, w, stack, stackptr)) {
 
         if (frm[y*w + x]!=nc) {
-            frm[y*w + x] = nc;
+            frm[y*w + x] = (BYTE) nc;
             (*sz)++;
         }
 
@@ -120,7 +120,7 @@ void flood_fill4_sz(int x, int y, int nc, int oc, int w, int h, BYTE *frm, int *
                 if(!push(x-1, y, w, stack, stackptr, MAX_BLK_RESOLUTION)) return;
             }
         }
-        if (y+1 <h) { 
+        if (y+1 <h) {
             if (frm[(y+1)*w + x]==oc && frm[(y+1)*w + x]!=nc) {
                 if(!push(x, y+1, w, stack, stackptr, MAX_BLK_RESOLUTION)) return;
             }
@@ -134,10 +134,10 @@ void flood_fill4_sz(int x, int y, int nc, int oc, int w, int h, BYTE *frm, int *
 }
 
 //Non-recursively flood-fills a 4-connected component in a block and collects the component's size and bounding box
-void flood_fill4_sz_bb(int x, int y, int nc, int oc, int w, int h, BYTE *frm, int *sz, int *min_x, int *max_x, int *min_y, int *max_y, int stack[], int& stackptr, int MAX_BLK_RESOLUTION) 
+void flood_fill4_sz_bb(int x, int y, int nc, int oc, int w, int h, BYTE *frm, int *sz, int *min_x, int *max_x, int *min_y, int *max_y, int stack[], int& stackptr, int MAX_BLK_RESOLUTION)
 {
     (*sz) = 0;
-    (*max_x) = (*max_y) = 0; 
+    (*max_x) = (*max_y) = 0;
     (*min_x) = (*min_y) = INT_MAX;
     if(nc == oc) return; //avoid infinite loop
     empty_stack(w, stack, stackptr);
@@ -147,7 +147,7 @@ void flood_fill4_sz_bb(int x, int y, int nc, int oc, int w, int h, BYTE *frm, in
     while(pop(&x, &y, w, stack, stackptr)) {
 
         if (frm[y*w + x]!=nc) {
-            frm[y*w + x] = nc;
+            frm[y*w + x] = (BYTE) nc;
             (*sz)++;
             if ((*min_x) > x) (*min_x) = x;
             if ((*max_x) < x) (*max_x) = x;
@@ -166,7 +166,7 @@ void flood_fill4_sz_bb(int x, int y, int nc, int oc, int w, int h, BYTE *frm, in
                 if(!push(x-1, y, w, stack, stackptr, MAX_BLK_RESOLUTION)) return;
             }
         }
-        if (y+1 <h) { 
+        if (y+1 <h) {
             if (frm[(y+1)*w + x]==oc && frm[(y+1)*w + x]!=nc) {
                 if(!push(x, y+1, w, stack, stackptr, MAX_BLK_RESOLUTION)) return;
             }
@@ -180,12 +180,12 @@ void flood_fill4_sz_bb(int x, int y, int nc, int oc, int w, int h, BYTE *frm, in
 }
 
 //Non-recursively flood-fills a 4-connected component in a block and collects the component's size, bounding box and sums of elements from the masks
-void flood_fill4_sz_bb_sum(int x, int y, int nc, int oc, int w, int h, BYTE *frm, int *sz, int *min_x, int *max_x, int *min_y, int *max_y, BYTE *mask, int *sum, BYTE *mask2, int *sum2, int stack[], int& stackptr, int MAX_BLK_RESOLUTION) 
+void flood_fill4_sz_bb_sum(int x, int y, int nc, int oc, int w, int h, BYTE *frm, int *sz, int *min_x, int *max_x, int *min_y, int *max_y, BYTE *mask, int *sum, BYTE *mask2, int *sum2, int stack[], int& stackptr, int MAX_BLK_RESOLUTION)
 {
     (*sz) = 0;
     (*sum) = 0;
     (*sum2) = 0;
-    (*max_x) = (*max_y) = 0; 
+    (*max_x) = (*max_y) = 0;
     (*min_x) = (*min_y) = INT_MAX;
     if(nc == oc) return; //avoid infinite loop
     empty_stack(w, stack, stackptr);
@@ -195,7 +195,7 @@ void flood_fill4_sz_bb_sum(int x, int y, int nc, int oc, int w, int h, BYTE *frm
     while(pop(&x, &y, w, stack, stackptr)) {
 
         if (frm[y*w + x]!=nc) {
-            frm[y*w + x] = nc;
+            frm[y*w + x] = (BYTE) nc;
             (*sz)++;
             (*sum)+= mask[y*w + x];
             (*sum2)+= mask2[y*w + x];
@@ -216,7 +216,7 @@ void flood_fill4_sz_bb_sum(int x, int y, int nc, int oc, int w, int h, BYTE *frm
                 if(!push(x-1, y, w, stack, stackptr, MAX_BLK_RESOLUTION)) return;
             }
         }
-        if (y+1 <h) { 
+        if (y+1 <h) {
             if (frm[(y+1)*w + x]==oc && frm[(y+1)*w + x]!=nc) {
                 if(!push(x, y+1, w, stack, stackptr, MAX_BLK_RESOLUTION)) return;
             }
